restructure, add some externally observable status/progress info
